requirements.txt
pytest==7.4.3
selene==2.0.0rc4
allure-pytest
Appium-Python-Client